revert: remove hosted run intent execution
Reverts the complete change introduced by #929 at the repository owner's request.
