+++ Cargo.toml
@@ -6,7 +6,6 @@ members = [
   "crates/openengine-cluster-client",
   "crates/openengine-cluster-testkit",
   "crates/openengine-cluster-testkit/tests/external-backend-fixture",
-  "docker/zeroshot-oecp",
   "zeroshot-rust",
 ]
 
@@ -18,7 +17,6 @@ repository = "https://github.com/the-open-engine/zeroshot"
 
 [workspace.dependencies]
 async-trait = "0.1.89"
-axum = { version = "0.8.4", features = ["json"] }
 fs2 = "0.4.3"
 getrandom = "0.3.4"
 libc = "0.2.177"
@@ -29,7 +27,7 @@ serde = { version = "1.0.228", features = ["derive"] }
 serde_json = "1.0.145"
 sha2 = "0.10.9"
 thiserror = "2.0.17"
-tokio = { version = "1.48.0", features = ["fs", "io-std", "io-util", "macros", "net", "process", "rt-multi-thread", "signal", "sync", "time"] }
+tokio = { version = "1.48.0", features = ["fs", "io-std", "io-util", "macros", "net", "process", "rt-multi-thread", "sync", "time"] }
 tokio-util = { version = "0.7", features = ["codec"] }
 tokio-stream = "0.1"
 tokio-tungstenite = "0.26"
