+++ lib/cluster-worker/engine-adapter.js
@@ -1,7 +1,6 @@
 'use strict';
 
 const { cloneJson } = require('./object-utils');
-const { inputFromRequest } = require('./engine-input');
 
 const TERMINAL_TOPICS = new Set(['CLUSTER_COMPLETE', 'CLUSTER_FAILED']);
 
@@ -12,14 +11,9 @@ function messageKey(message) {
 function terminalEventFromMessage(message) {
   const data = message.content?.data || {};
   if (message.topic === 'CLUSTER_COMPLETE') {
-    let prSummary = null;
-    if (typeof data.pr_url === 'string' && data.pr_url) {
-      const number = Number.isSafeInteger(data.pr_number) ? ` #${data.pr_number}` : '';
-      prSummary = `Pull request${number} created: ${data.pr_url}`;
-    }
     return {
       type: 'complete',
-      summary: prSummary || data.reason || message.content?.text,
+      summary: data.reason || message.content?.text,
       ...(Object.prototype.hasOwnProperty.call(data, 'result') ? { result: data.result } : {}),
       ...(Object.prototype.hasOwnProperty.call(data, 'artifacts')
         ? { artifacts: data.artifacts }
@@ -31,6 +25,17 @@ function terminalEventFromMessage(message) {
     summary: data.reason || message.content?.text,
     code: data.code,
     reason: data.workerReason,
+  };
+}
+
+function inputFromRequest(request, artifactManifest) {
+  if (request.source === 'issue') return { issue: request.issue };
+  if (request.source === 'prompt') return { text: request.prompt };
+  return {
+    text:
+      'Execute the task described by this registry-staged, byte-free artifact manifest. ' +
+      'Do not request interactive input.\n' +
+      JSON.stringify(artifactManifest),
   };
 }
 
@@ -233,16 +238,16 @@ class CurrentEngineAdapter {
     this.stopPromise ||= this.stopResource();
     return this.stopPromise;
   }
-
   async waitForAllocatedCluster() {
     let cluster = this.orchestrator.getCluster(this.resource.clusterId);
     while (!cluster && !this.startSettled && !this.closed) {
-      await new Promise((resolve) => setTimeout(resolve, 10));
+      await new Promise((resolve) => {
+        setTimeout(resolve, 10);
+      });
       cluster = this.orchestrator.getCluster(this.resource.clusterId);
     }
     return cluster;
   }
-
   stopAllocatedCluster(cluster = this.orchestrator.getCluster(this.resource.clusterId)) {
     if (!cluster) return Promise.resolve(false);
     if (cluster.state === 'stopped' || cluster.state === 'killed') return Promise.resolve(true);
@@ -251,7 +256,6 @@ class CurrentEngineAdapter {
       .then(() => true);
     return this.allocatedStopPromise;
   }
-
   async stopResource() {
     this.cleanupPromise ||= this.waitForAllocatedCluster()
       .then((cluster) => this.stopAllocatedCluster(cluster))
