+++ lib/cluster-worker/profiles.js
@@ -24,44 +24,6 @@ const DEFAULT_PROVIDER_PROFILES = Object.freeze({
   }),
 });
 
-const HOSTED_CODEX_OPENROUTER_PROFILE = 'provider.codex-openrouter@1';
-const HOSTED_CODEX_OPENROUTER_PR_PROFILE = 'provider.codex-openrouter-pr@1';
-const DEFAULT_HOSTED_MODEL = 'openai/gpt-5.4';
-
-function providerProfilesFromEnvironment(environment = process.env) {
-  if (environment.ZEROSHOT_HOSTED_CODEX_OPENROUTER !== '1') {
-    return DEFAULT_PROVIDER_PROFILES;
-  }
-  const model = environment.ZEROSHOT_HOSTED_MODEL || DEFAULT_HOSTED_MODEL;
-  const levelOverrides = Object.freeze({
-    level1: Object.freeze({ model, reasoningEffort: 'medium' }),
-    level2: Object.freeze({ model, reasoningEffort: 'high' }),
-    level3: Object.freeze({ model, reasoningEffort: 'xhigh' }),
-  });
-  const hostedSettings = {
-    templateParams: { task_type: 'TASK' },
-    providerSettings: {
-      codex: {
-        defaultLevel: 'level2',
-        levelOverrides,
-      },
-    },
-  };
-  return deepFreeze({
-    ...DEFAULT_PROVIDER_PROFILES,
-    [HOSTED_CODEX_OPENROUTER_PROFILE]: {
-      configName: 'base-templates/single-worker',
-      providerOverride: 'codex',
-      settings: hostedSettings,
-    },
-    [HOSTED_CODEX_OPENROUTER_PR_PROFILE]: {
-      configName: 'base-templates/worker-validator',
-      providerOverride: 'codex',
-      settings: hostedSettings,
-    },
-  });
-}
-
 function own(map, key) {
   return Object.prototype.hasOwnProperty.call(map, key) ? map[key] : null;
 }
@@ -76,7 +38,7 @@ function validateBounds(bounds) {
 
 function createDeploymentProfileRegistry(options = {}) {
   const isolationProfiles = options.isolationProfiles || DEFAULT_ISOLATION_PROFILES;
-  const providerProfiles = options.providerProfiles || providerProfilesFromEnvironment();
+  const providerProfiles = options.providerProfiles || DEFAULT_PROVIDER_PROFILES;
   const defaultBounds = { ...DEFAULT_BOUNDS, ...(options.bounds || {}) };
   validateBounds(defaultBounds);
   const bounds = deepFreeze(defaultBounds);
@@ -115,8 +77,5 @@ module.exports = {
   DEFAULT_BOUNDS,
   DEFAULT_ISOLATION_PROFILES,
   DEFAULT_PROVIDER_PROFILES,
-  HOSTED_CODEX_OPENROUTER_PROFILE,
-  HOSTED_CODEX_OPENROUTER_PR_PROFILE,
   createDeploymentProfileRegistry,
-  providerProfilesFromEnvironment,
 };
