+++ lib/start-cluster.js
@@ -176,17 +176,17 @@ function applyProviderOverrideToConfig(config, providerOverride, settings) {
   console.log(chalk.dim(`Provider override: ${providerOverride} (all agents)`));
 }
 
-function resolveParameterizedConfigFile(config, templateParams = {}) {
+function resolveParameterizedConfigFile(config) {
   if (!config?.params || Object.keys(config.params).length === 0) {
     return config;
   }
 
   const resolver = new TemplateResolver(path.join(PACKAGE_ROOT, 'cluster-templates'));
-  return resolver.resolveTemplate(config, templateParams);
+  return resolver.resolveTemplate(config, {});
 }
 
 function prepareClusterConfig(config, settings = {}, providerOverride) {
-  const prepared = resolveParameterizedConfigFile(config, settings.templateParams || {});
+  const prepared = resolveParameterizedConfigFile(config);
   ensureConfigProviderDefaults(prepared, settings);
   if (providerOverride) {
     applyProviderOverrideToConfig(prepared, providerOverride, settings);
