+++ zeroshot-rust/src/lib.rs
@@ -103,4 +103,3 @@ where
         Default::default(),
     )
 }
-pub mod hosted_oecp;
